+++ types/src/texter.d.ts
@@ -1,0 +1,7 @@
+export const T: {
+	invalidStructChar: (value: any) => string;
+	invalidConstructorRaw: (value: any) => string;
+	invalidCursor: (value: any) => string;
+	invalidSzie: (value: any) => string;
+	invalidOptions: (value: any) => string;
+};